Refactor manual hashing control (fully back-compatible)
Hashing is still done on a data level as before (with arb data-hash-fn);
difference is that {:timbre/hash _} arg0 support is now baked in to core
logging impl. rather than just the default data hash fn.

Effect: a new `:?hash-arg` is now provided as part of core data and will
be available to any data-hash-fns.

New implementation is also cleaner + faster.