Update example-config in Readme
The appender name :example-println-appender was a bit confusing and I had to
dig into the source code to find out the default name was :println

I also noticed some of the options in the Readme were different from the
defaults in the code (:rate-limit, for example) and the Readme states the config
shown is "Timbre's default config"

This pull request reproduces exactly what is in the source code at this moment,
at the expense of removing some useful options from the example. Maybe those
could be added as a separate example in the Readme?