[#173] Experimental: re-enable auto binding conveyance for profiling
Have also modified the capturing strategy to eliminate contention
during compaction/merge.

So this is slower than the previous non-conveying implementation
(~2-4x the overhead):
  1. Clojure's ^:dynamic bindings are just inherently sluggish.
  2. Possible thread contention means that we need to switch to
     an immutable accumulator + atom.

Question is: is that actually a significant difference in practice?
We're talking about ~200-400ms overhead  per million+ calls.

If we're profiling something in production, it's presumably
because it's worth profiling (i.e. relatively expensive). And if
it's relatively expensive, then this overhead will very likely be
insignificant (incl. the nil *pdata_* => noop case).

So from a performance pov it might be reasonable to just auto-convey,
even if it's only rarely useful.

But ignoring performance: is the (~rare) usefulness of auto conveyance
worth the extra conceptual complexity / possible confusion?

Will take some time to think on this. Feedback welcome.