[#350] Low-level `log!` fn should support runtime arguments
The `log!` macro was intended as a fairly low-level util for tooling
authors. For example, to allow the convenient creation of logging wrappers.

But before this commit, `log!` unnecessarily required that its `args`
input be a *compile-time* sequential type.

That's fine for writing macro wrappers, but problematic for fn wrappers.

So this commit removes that requirement, and explicitly supports the case
where `args` is a symbol at compile-time - i.e. the case of a fn wrapper:

  (fn wrapper [& args] (timbre/log! :info :p args))