[new] [#354] Make `callsite-id` in `log!` macro deterministic for Clojure (@DerGuteMoritz)
Since the non-deterministic `rand` was used as part of generating
`callsite-id` in the `log!` macro at compile time, the compiler would
produce non-deterministic bytecode as a result. This thwarts
reproducible builds.

Instead, use a RNG with a fixed seed. Unfortunately, both the JS
standard library and Google Closure's library don't seem to provide a
similar facility, so it's only implemented for Clojure.