[mod] Change default top-level `*config*` :min-level from :report -> :trace
A question from a user just highlighted that the current default level behaviour
is unintuitive.

If `*config*` doesn't have any `:min-level` (rare!), that should imply that no minimum
level exists (i.e. any level should be fine). But instead the current behaviour is
flipped (no `:min-level` restricts all but `:report` level logs).

Note that the appender's `:min-level` already defaults to the (intuitively sensible)
`:trace`, so the current behaviour is not only surprising but also inconsistent.

This is a BREAKING CHANGE but should hopefully affect few folks since it's so rare
for anyone to have a `*config*` without an explicit `:min-level`.